type replacement

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(double weight, double height) {
+    public int calculate(double weight, double height) {
         double bmiIndex = weight / Math.pow(height, 2);
         return (int) bmiIndex;
     }
